requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
crispy-bootstrap4==2026.2
defusedxml==0.7.1
dj-database-url==3.1.2
Django==5.2.12
django-allauth==65.15.0
django-countries==8.2.0
django-crispy-forms==2.6
django-debug-toolbar==6.2.0
gunicorn==25.1.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
python3-openid==3.2.0
pytz==2026.1.post1
razorpay==2.0.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
